+++ src/articles/how-stoicism-helps.md
