+++ src/articles/lessons-from-steve-jobs.md
